api/requirements.txt
djangorestframework==3.5.3
django==1.11.4
django-cors-headers==2.1.0
